Changed the method for converting the flow-equivalent tree into a matrix. This appears to be a serious bottleneck.

+++ common/context.cpp
@@ -121,7 +121,7 @@ namespace multistateTurnip
 		std::size_t nVertices = boost::num_vertices(*unorderedGraph);
 		int minVertexIndex = *std::min_element(interestVertices->begin(), interestVertices->end());
 		int maxVertexIndex = *std::max_element(interestVertices->begin(), interestVertices->end());
-		if(minVertexIndex < 0 || maxVertexIndex >= nVertices)
+		if(minVertexIndex < 0 || maxVertexIndex >= (int)nVertices)
 		{
 			throw std::runtime_error("Input interestVertices was out of range");
 		}
